Added edit screen for recipe steps. Made a few minor responsive changes. Added all new recipe steps for risotto recipe,

+++ Data.Models/Models/RecipeStep.cs
@@ -18,6 +18,6 @@
         // When we create a cooking recipe step, keep track of timings
         public bool HasTimer { get; set; }
 
-        public int TimerDuration { get; set; }
+        public decimal TimerDuration { get; set; } // seconds
     }
 }
